requirements.txt
Django==3.1
djangorestframework==3.11
psycopg2-binary>=2.8
django-cors-headers
django-redis
Faker
coverage